+++ core/services/chainlink/config.go
@@ -101,18 +101,22 @@ func (s *Secrets) Validate() error {
 	return config.Validate(s)
 }
 
-// setOverrides overrides fields with values from ENV vars and password files.
-func (s *Secrets) setOverrides(keystorePasswordFileName, vrfPasswordFileName *string) error {
-	// Override DB and Explorer secrets from ENV vars, if present
+// setEnv overrides fields from ENV vars, if present.
+func (s *Secrets) setEnv() error {
 	if dbURL := config.EnvDatabaseURL.Get(); dbURL != "" {
+		s.Database.URL = new(models.SecretURL)
 		if err := s.Database.URL.UnmarshalText([]byte(dbURL)); err != nil {
 			return err
 		}
 	}
 	if dbBackupUrl := config.EnvDatabaseBackupURL.Get(); dbBackupUrl != "" {
+		s.Database.BackupURL = new(models.SecretURL)
 		if err := s.Database.BackupURL.UnmarshalText([]byte(dbBackupUrl)); err != nil {
 			return err
 		}
+	}
+	if config.EnvDatabaseAllowSimplePasswords.IsTrue() {
+		s.Database.AllowSimplePasswords = true
 	}
 	if explorerKey := config.EnvExplorerAccessKey.Get(); explorerKey != "" {
 		s.Explorer.AccessKey = &explorerKey
@@ -129,8 +133,11 @@ func (s *Secrets) setOverrides(keystorePasswordFileName, vrfPasswordFileName *st
 	if pyroscopeAuthToken := config.EnvPyroscopeAuthToken.Get(); pyroscopeAuthToken != "" {
 		s.Pyroscope.AuthToken = &pyroscopeAuthToken
 	}
+	return nil
+}
 
-	// Override Keystore and VRF passwords from corresponding files, if present
+// setPasswords overrides Keystore and VRF passwords, if present.
+func (s *Secrets) setPasswords(keystorePasswordFileName, vrfPasswordFileName *string) error {
 	if keystorePasswordFileName != nil {
 		keystorePwd, err := utils.PasswordFromFile(*keystorePasswordFileName)
 		if err != nil {